Xcode/GCC warning fixes.

+++ interface/src/devices/SixenseManager.cpp
@@ -60,7 +60,7 @@ void SixenseManager::update(float deltaTime) {
         // Either find a palm matching the sixense controller, or make a new one
         PalmData* palm;
         bool foundHand = false;
-        for (int j = 0; j < hand->getNumPalms(); j++) {
+        for (size_t j = 0; j < hand->getNumPalms(); j++) {
             if (hand->getPalms()[j].getSixenseID() == data.controller_index) {
                 palm = &(hand->getPalms()[j]);
                 foundHand = true;
@@ -128,7 +128,7 @@ void SixenseManager::update(float deltaTime) {
     }
     
     // if the controllers haven't been moved in a while, disable
-    const int MOVEMENT_DISABLE_DURATION = 30 * 1000 * 1000;
+    const unsigned int MOVEMENT_DISABLE_DURATION = 30 * 1000 * 1000;
     if (usecTimestampNow() - _lastMovement > MOVEMENT_DISABLE_DURATION) {
         for (vector<PalmData>::iterator it = hand->getPalms().begin(); it != hand->getPalms().end(); it++) {
             it->setActive(false);
